# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Pegadaian",
        "description": "created by nur arifin",
        "contact": {
            "email": "arifin@lenna.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://pegadaian.test",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authenticate User"
                ],
                "summary": "Register User",
                "description": "Untuk Register User",
                "operationId": "Register User",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "name": "string",
                                            "email": "string",
                                            "access_token": "token",
                                            "token_type": "bearer"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authenticate User"
                ],
                "summary": "login User",
                "description": "Untuk login User",
                "operationId": "login User",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "arifin@lenna.ai",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "name": "string",
                                            "email": "string",
                                            "roles": [
                                                {
                                                    "name": "admin",
                                                    "permissions": [
                                                        {
                                                            "name": "create_user"
                                                        },
                                                        {
                                                            "name": "update_user"
                                                        },
                                                        {
                                                            "name": "read_user"
                                                        },
                                                        {
                                                            "name": "delete_user"
                                                        }
                                                    ]
                                                }
                                            ],
                                            "access_token": "token",
                                            "token_type": "bearer"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Authenticate User"
                ],
                "summary": "current User",
                "description": "current User",
                "operationId": "current User",
                "responses": {
                    "202": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "id": 10,
                                            "name": "arifin",
                                            "email": "ssarifin@gmails.scoma",
                                            "email_verified_at": null,
                                            "created_at": "2023-11-24T15:14:25.000000Z",
                                            "updated_at": "2023-11-24T15:14:25.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authenticate User"
                ],
                "summary": "logout User",
                "description": "Untuk logout User",
                "operationId": "logout User",
                "responses": {
                    "204": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Successfully logged out"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authenticate User"
                ],
                "summary": "Refresh User",
                "description": "Untuk Refresh User",
                "operationId": "Refresh User",
                "responses": {
                    "202": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "name": "string",
                                            "email": "string",
                                            "access_token": "token",
                                            "token_type": "bearer"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}